refactor(root): Reorganized OS specific things

+++ install-scripts/install.sh
@@ -5,49 +5,73 @@
 # - Open Tmux and install tmux plugins (^a + I)
 # - Add Nord theme to Slack
 
-#### Mac Install
+##################################################
+#  Mac Install
+if [ $OSTYPE =~ darwin* ]; then
 
 # Install Homebrew
-# /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-
+/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 # Install tools
-# brew update && brew upgrade
-# brew install coreutils \
-#   alacritty \
-#   trash \
-#   zsh \
-#   tmux \
-#   neovim \
-#   yabai \
-#   skhd \
-#   fzf \
-#   ripgrep \
-#   z
-# brew tap homebrew/command-not-found
+brew update && brew upgrade
+brew install coreutils \
+  alacritty \
+  trash \
+  zsh \
+  tmux \
+  neovim \
+  yabai \
+  skhd \
+  fzf \
+  ripgrep \
+  z
+# Command-not-found functionality like Ubuntu's apt-get
+brew tap homebrew/command-not-found
 
-#### Arch Linux Install
+# Yabai settings
+[ -e ~/.config/yabai ] && rm -rf ~/.config/yabai
+mkdir -p ~/.config/yabai
+ln -sf $(realpath ./yabairc) ~/.config/yabai/yabairc
+# SKHD settings
+[ -e ~/.config/skhd ] && rm -rf ~/.config/skhd
+mkdir -p ~/.config/skhd
+ln -sf $(realpath ./skhdrc) ~/.config/skhd/skhdrc
 
-# pacman -Syu alacritty \
-#   zsh \
-#   tmux \
-#   neovim \
-#   fzf \
-#   ripgrep \
-#   z \
-#   lm-sensors \
-#   free
+##################################################
+# Debian-based Install
+elif [ $OSTYPE =~ linux* ] && command -v apt-get &> /dev/null; then
 
-#### Debian-based Install
+apt-get update && apt-get upgrade
+apt-get install alacritty \
+  zsh \
+  tmux \
+  neovim \
+  fzf \
+  ripgrep \
+  z \
+  lm-sensors \
 
-# apt-get update && apt-get upgrade
-# apt-get install alacritty \
-#   zsh \
-#   tmux \
-#   neovim \
-#   fzf \
-#   ripgrep \
-#   z \
-#   lm-sensors \
+##################################################
+# Arch Linux Install
+elif [ $OSTYPE =~ linux* ] && command -v pacman &> /dev/null; then
+
+pacman -Syu alacritty \
+  zsh \
+  tmux \
+  neovim \
+  fzf \
+  ripgrep \
+  z \
+  lm-sensors \
+  free
+
+# Sway (tiling window manager)
+[ -e ~/.config/sway ] && rm -rf ~/.config/sway
+ln -sfn $(realpath ./sway) ~/.config/sway
+
+fi
+
+##################################################
+# The rest is platform agnostic
 
 # I assume git is already installed
 # Set git config
@@ -75,6 +99,7 @@ rm lf-linux-amd64.tar.gz
 
 # Install NVM
 curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash
+nvm install --lts
 
 # Add custom plugins to oh-my-zsh directory
 git clone https://github.com/zsh-users/zsh-autosuggestions $ZSH_CUSTOM/plugins/zsh-autosuggestions
@@ -84,10 +109,12 @@ git clone https://github.com/nyquase/vi-mode $ZSH_CUSTOM/plugins/vi-mode
 # Installs tmux plugin manager
 git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 
-# TODO move these into .config
-rm ~/.zshrc
+# ZSH settings
+[ -e ~/.zshrc ] && rm ~/.zshrc
 ln -s $(realpath ./zshrc) ~/.zshrc
-rm ~/.tmux.conf
+
+# Tmux Settings
+[ -e ~/.tmux.conf ] && rm ~/.tmux.conf
 ln -s $(realpath ./tmux.conf) ~/.tmux.conf
 
 # Symlink Neovim's config
@@ -95,7 +122,7 @@ mkdir -p ~/.config/nvim
 ln -s $(realpath ./vimrc) ~/.config/nvim/init.vim
 
 # Create UltiSnips in .vim folder
-rm -rf ~/.config/coc
+[ -e ~/.config/coc ] && rm -rf ~/.config/coc
 mkdir -p ~/.config/coc
 ln -sfn $(realpath ./ultisnips) ~/.config/coc/ultisnips
 ln -sf $(realpath ./coc-settings.json) ~/.config/nvim/coc-settings.json
@@ -105,36 +132,19 @@ mkdir -p ~/.config/nvim/autoload/floaterm/wrapper
 ln -s $(realpath ./fzf.vim) ~/.config/nvim/autoload/floaterm/wrapper/fzf.vim
 
 # Symlink Alacritty folder for both Mac & Linux
-rm -rf ~/.config/alacritty
+[ -e ~/.config/alacritty ] && rm -rf ~/.config/alacritty
 mkdir -p ~/.config/alacritty
 ln -s $(realpath ./alacritty.yml) ~/.config/alacritty/alacritty.yml
 
 # Symlink lf config
-rm -rf ~/.config/lf
+[ -e ~/.config/lf ] && rm -rf ~/.config/lf
 ln -sfn $(realpath ./lf) ~/.config/lf
 
 # Symlink Starship configuration
-rm ~/.config/starship.toml
+[ -e ~/.config/starship.toml ] && rm ~/.config/starship.toml
 mkdir -p ~/.config/starship
 ln -s $(realpath ./starship.toml) ~/.config/starship/toml
 
 # Install vim-plug (should work on Mac & Linux)
 # Neovim
 curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-
-#### Mac-specific settings
-
-# rm -rf ~/.config/yabai
-# mkdir -p ~/.config/yabai
-# rm -rf ~/.config/skhd
-# mkdir -p ~/.config/skhd
-# ln -sf $(realpath ./yabairc) ~/.config/yabai/yabairc
-# ln -sf $(realpath ./skhdrc) ~/.config/skhd/skhdrc
-
-#### Linux-specific settings
-
-# Symlinks for Sway
-# rm -rf ~/.config/sway
-# mkdir -p ~/.config/sway
-# ln -sf $(pwd)/swayconfig ${HOME}/.config/sway/config
-# ln -sf $(pwd)/swaybar.sh ${HOME}/.config/sway/swaybar.sh
